# services/api-gateway/api/openapi.yaml
openapi: 3.0.3
info:
  title: Unified Task Manager API
  version: 0.1.0
  description: |
    Public HTTP API exposed through the API Gateway (`/api` prefix on the edge).
    Nginx strips `/api` before proxying; paths below are relative to that prefix.

    | Prefix | Service |
    |--------|---------|
    | `/v1/auth`, `/v1/users`, `/v1/teams`, `/v1/permissions` | user-service |
    | `/v1/tasks`, `/v1/task-columns`, `/v1/boards` | task-service |
    | `/v1/chats` | chat-service |
    | `/v1/projects`, `/v1/tasks/.../deadline-notification` | notification-service |
    | `/v1/file-environments` | file-service |
    | `/api/tasks/assistant`, `/task_name_description` | ml-service |
    | `/calls` | sfu-service (WebRTC signaling at `/ws/calls` on the edge) |

    **Auth:** `Authorization: Bearer <accessToken>` except public auth routes.
    **Team scope:** send `X-Team-Id` for task/column APIs (gateway also accepts `teamId` query on board WebSocket).
    **Optimistic concurrency:** task/column PATCH accepts `If-Match` (ETag from GET/PATCH); `412` returns `VERSION_CONFLICT` and `current` entity.
servers:
  - url: /api
    description: API Gateway (browser clients use this base URL)
...
security:
  - bearerAuth: []
paths:
  /v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/users:
    get:
      tags: [Users]
      summary: List users
      description: Intended for admin or internal use via gateway policy.
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /v1/teams:
    get:
      tags: [Teams]
      summary: List teams for current user
      operationId: listTeams
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Teams]
      summary: Create team
      operationId: createTeam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeamRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/permissions/check:
    post:
      tags: [Permissions]
      summary: Check project permission for current user
      operationId: checkPermission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionCheckRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionCheckResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/tasks:
    get:
      tags: [Tasks]
      summary: List tasks in project
      operationId: listTasks
      parameters:
        - $ref: '#/components/parameters/TeamIdHeader'
        - $ref: '#/components/parameters/ProjectIdQuery'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          headers:
            ETag:
              description: Not set on list responses
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Tasks]
      summary: Create task
      operationId: createTask
      parameters:
        - $ref: '#/components/parameters/TeamIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /v1/task-columns:
    get:
      tags: [Tasks]
      summary: List columns for project
      operationId: listTaskColumns
      parameters:
        - $ref: '#/components/parameters/TeamIdHeader'
        - $ref: '#/components/parameters/ProjectIdQuery'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskColumnListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Tasks]
      summary: Create column
      operationId: createTaskColumn
      parameters:
        - $ref: '#/components/parameters/TeamIdHeader'
        - $ref: '#/components/parameters/ProjectIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
      responses:
        '201':
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskColumn'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/tasks/assistant:
    post:
      tags: [ML]
      summary: Run task/file assistant (single response)
      operationId: runTaskAssistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskFileAgentRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskFileAgentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /task_name_description:
    post:
      tags: [ML]
      summary: Generate task title and description from raw text
      operationId: generateTaskNameDescription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskNameDescriptionRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskNameDescriptionResponse'
  /calls:
    post:
      tags: [Calls]
      summary: Create call session
      operationId: createCall
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallSession'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    bearerQuery:
      type: apiKey
      in: query
      name: token
      description: Bearer access token in query (WebSocket clients)
  parameters:
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    TeamIdHeader:
      name: X-Team-Id
      in: header
      required: true
      schema:
        type: string
      description: Active team for task/column APIs
...
    ProjectIdQuery:
      name: projectId
      in: query
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Not enough permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    User:
      type: object
      required:
        - id
        - email
        - name
        - role
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
          minLength: 1
          maxLength: 120
        role:
          type: string
          enum: [user, manager, admin]
        status:
          type: string
          enum: [active, inactive, suspended]
        tag:
          type: string
        bio:
          type: string
        githubUrl:
          type: string
        linkedInUrl:
          type: string
        telegram:
          type: string
        websiteUrl:
          type: string
        secondaryEmail:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
        name:
          type: string
          minLength: 1
          maxLength: 120
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    TokenPair:
      type: object
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: Access token lifetime in seconds
    AuthResponse:
      type: object
      required: [user, tokens]
      properties:
        user:
          $ref: '#/components/schemas/User'
        tokens:
          $ref: '#/components/schemas/TokenPair'
...
    UserListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/User'
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
        offset:
          type: integer
          minimum: 0
    ErrorResponse:
      type: object
      required: [message]
      properties:
        code:
          type: string
          description: Present on user-service errors (e.g. UNAUTHORIZED)
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    Team:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TeamListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Team'
    CreateTeamRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
...
    PermissionCheckRequest:
      type: object
      required: [projectId, permission]
      properties:
        teamId: { type: string }
        projectId: { type: string }
        permission: { type: string }
    PermissionCheckResult:
      type: object
      required: [allowed]
      properties:
        allowed: { type: boolean }
        source: { type: string, enum: [team, project] }
        teamRoleKey: { type: string }
        projectRoleKey: { type: string }
        matched:
          type: array
          items: { type: string }
    Task:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        status: { type: string, enum: [todo, in_progress, done] }
        priority: { type: string, enum: [low, medium, high] }
        dueAt: { type: string, format: date-time, nullable: true }
        completedAt: { type: string, format: date-time, nullable: true }
        completedBy: { type: string }
        createdBy: { type: string }
        assigneeUserId: { type: string }
        assigneeName: { type: string }
        teamId: { type: string }
        projectId: { type: string }
        unreadComments: { type: integer }
        version: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TaskListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
    CreateTaskRequest:
      type: object
      required: [title, projectId]
      properties:
        title: { type: string }
        description: { type: string }
        status: { type: string, enum: [todo, in_progress, done] }
        priority: { type: string, enum: [low, medium, high] }
        dueAt: { type: string, format: date-time }
        assigneeUserId: { type: string }
        assigneeName: { type: string }
        projectId: { type: string }
...
    TaskColumn:
      type: object
      properties:
        id: { type: string }
        teamId: { type: string }
        projectId: { type: string }
        title: { type: string }
        position: { type: integer }
        version: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TaskColumnListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TaskColumn'
...
    TaskFileAgentRequest:
      type: object
      required: [request_id, message, team_id, project_id]
      properties:
        request_id: { type: string }
        message: { type: string }
        team_id: { type: string }
        project_id: { type: string }
        max_iterations: { type: integer, minimum: 1, maximum: 12, default: 8 }
        include_capabilities: { type: boolean, default: true }
    TaskFileAgentResponse:
      type: object
      properties:
        request_id: { type: string }
        answer: { type: string }
        succeeded: { type: boolean }
        tool_calls: { type: array, items: { type: object } }
        capabilities: { type: array, items: { type: object } }
    TaskNameDescriptionRequest:
      type: object
      required: [request_id, raw_task_description]
      properties:
        request_id: { type: string }
        raw_task_description: { type: string }
        current_task_name: { type: string }
    TaskNameDescriptionResponse:
      type: object
      properties:
        request_id: { type: string }
        raw_task_description: { type: string }
        task_name: { type: string }
        task_description: { type: string }
    CallSession:
      type: object
      properties:
        id: { type: string }
        initiator_id: { type: string }
        created_at: { type: string, format: date-time }
        ended_at: { type: string, format: date-time }
        status: { type: string, enum: [active, ended] }
        ice_servers:
          type: array
          items:
            type: object
            additionalProperties: { type: string }